requirements.txt
flask
requests
tvdsb-student-api>=1.6